Changed directKinematics return value to kinematicMatrix

+++ Robotic and Navigation in Medicine/src/Robot/UR5.cpp
@@ -19,7 +19,7 @@ bool UR5::connectToRobot(char* ip, int port){
 	return true;
 }
 
-bool UR5::setJoints(std::array<double,6> &angles) {
+bool UR5::setJoints(std::array<float,6> &angles) {
 	char jointString[100];
 
 	sprintf(jointString, "MovePTPJoints %f %f %f %f %f %f ", angles.at(0), angles.at(1), angles.at(2), angles.at(3), angles.at(4), angles.at(5));
@@ -29,13 +29,13 @@ bool UR5::setJoints(std::array<double,6> &angles) {
 
 }
 
-    std::array<double,6>& UR5::getJoints(char* mode) {
-	std::array<double, 6> ret;
+    std::array<float,6>& UR5::getJoints(char* mode) {
+	std::array<float, 6> ret;
 	double test = 1;
 	const char* jointString;
 	jointString = tcp_client_->command("GetPositionJoints");
 	std::cout <<"string: " << jointString << std::endl;
-	sscanf(jointString, "%lf %lf %lf %lf %lf %lf", &ret[0], &ret[1], &ret[2], &ret[3], &ret[4], &ret[5]);
+	sscanf(jointString, "%f %f %f %f %f %f", &ret[0], &ret[1], &ret[2], &ret[3], &ret[4], &ret[5]);
 	
 	if (mode == "deg") {
 		//nothing to do because the values are already in degree
